fixed bugs

+++ src/main/java/com/tqs/plazzamarket/controllers/CategoryController.java
@@ -1,6 +1,7 @@
 package com.tqs.plazzamarket.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.tqs.plazzamarket.entities.Admin;
 import com.tqs.plazzamarket.entities.Category;
 import com.tqs.plazzamarket.repositories.CategoryRepository;
 import io.swagger.annotations.Api;
@@ -12,6 +13,8 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Map;
+
+import javax.servlet.http.HttpSession;
 
 @RestController
 @RequestMapping(path = "/api/category")
@@ -32,17 +35,19 @@ public class CategoryController {
 
     @ApiOperation(value = "Create new category")
     @PostMapping(path = "/addcategory", consumes = "application/json")
-    public ResponseEntity<Category> createCategory(@RequestBody Map<String, Object> categoryJSON) {
+    public ResponseEntity<Category> createCategory(@RequestBody Map<String, Object> categoryJSON, HttpSession session) {
+        if (session.getAttribute("user") == null || session.getAttribute("user").getClass() != Admin.class)
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         Category category = mapper.convertValue(categoryJSON, Category.class);
         category = categoryRepository.saveAndFlush(category);
-        System.out.println(category);
-        System.out.println(category.getName());
         return new ResponseEntity<>(category, HttpStatus.CREATED);
     }
 
     @ApiOperation(value = "Create new category")
     @DeleteMapping(path = "/delete/{category}")
-    public ResponseEntity<Boolean> deleteCategory(@PathVariable("category") String category) {
+    public ResponseEntity<Boolean> deleteCategory(@PathVariable("category") String category, HttpSession session) {
+        if (session.getAttribute("user") == null || session.getAttribute("user").getClass() != Admin.class)
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         if (!categoryRepository.findById(category).isPresent())
             return new ResponseEntity<>(false, HttpStatus.BAD_REQUEST);
         categoryRepository.deleteById(category);
